Rework the reaper thread to use a queue
The previous incarnation iterated over the whole map to find expired
resources. Now the reaper thread gets notifed via a TQueue to go check
the map to see if the resource is ready to be collected. If it has been
refreshed, the expiry will be out of date and it won't be collected.