fix issue that was causing ill-formed contexts
Ill formed contexts (not types) were causing a number of tests to fail.
During let rec checking / inference, we were using the same variable for both the marker and the first binding of the let rec. This resulted in contexts being ill-formed if the binding had a user-provided annotation but worked fine as long as the first binding was being inferred.

As for why we didn't notice this before, I think we have a typo in when we are calling minimize' (to eliminate needless let rec), so the let rec is not being removed and we are hitting a case that wasn't hit before.