Update Reference.showShort and Referent.showShort to only shorten the hash portion of the id.
Also renamed Reference's semiprivate constructors (Builtin_ and DerivedPrivate_) to their public names (Builtin and DerivedId)
This eliminates the completeness checking issue with pattern synonyms, and we were leaking the constructors anyway.

Reference cycle length is currently encoded in the Reference's textual form; we can and probably should drop it, as it is redundant, but we then need a codebase or list of known references when parsing a Reference from Text.