fixed the crash but...
It still produces some empty messages.  They occur when there's nothing to add to the codebase,
and everything is filtered out of the UF.  The `SlurpOutput` message renderer doesn't print anything
but some padding in this case.

So I think somehow ... oh don't actually filter out things that already exist in the codebase; just don't
include them in the check for expanding the selection list?