Trying to allow record fields to be polymorphic:
type Functor f = { map : forall a b . (a -> b) -> f a -> f b }

The current strategy of just inferring the type of

Functor.map.set g f = case f of
  Functor _ -> Functor g

...doesn't work, since this should be a higher rank function and inference
doesn't infer anything higher rank. Initially I was thinking it was as
simple as providing a signature, but the signature can't be trivially
computed from the type of the field.

It's doable, just needs more thought.