30 compile errors, but errors are now in tests, HandleInput, and command line
There are a bunch of places where we are constructing a `PrettyPrintEnv` from a `Names0` via `PrettyPrintEnv.fromNames0`. The function has gone away, replaced by a version that takes basically a pair of `Names0`, one for the current names, and one for the historical names. The historical names that are needed have to be constructed based on what the PrettyPrintEnv is being applied to, so some thought will need to go into figuring out how to upgrade each of the old usages of the `fromNames0` function.