output messages uses `PrettyPrintEnv` rather than `Names3.Names`.
This seems like the right move since `Names3.Names` needs hash length to convert to a pretty print env, which output messages won't have anyway of obtaining, and `Names0` is not sufficient for printing out definitions that may need hash qualification. Also observed that everywhere in OutputMessages, the `Names0` was getting converted to a `PrettyPrintEnv` and then the names thrown out.