down to 129 failures, pattern syntax treats symbols starting with uppercase as constructor references...
We made a half baked attempt to allow uppercase patttern variables, but making this work would require restructuring the parser - you need to know whether the pattern you are currently parsing is in head position or not. Didn't want to think through how to make this work without breaking things.