Don't assume RefNamed vars not to occur in input to Term.unhashComponent.
Term.unhashComponent used to make up some RefNamed variables and simply
assumed that they did not appear in the terms of the input component.
Such assumption is not captured by the type signature of `unhashComponent`.

To drop the assumption, freshen the made up RefNamed variables appropriately.

Add a test that previously would have failed.