Don't conflate multiple failure modes of Context.solve into Nothing.
It can fail for 3 different reasons, but only in one case should
typechecking continue. However, all three cases used to result in `Nothing`
and the caller then tried to reconstruct the reason why it failed.

Now one of them results in typechecking failure, another one results in
a compiler bug (which would previously be reported as type mismatch) and
the last one is returned as `Nothing`.