Make typechecking fail fast on adding non-reserved var to the context.
That is when a variable to the context for which there is no guarantee
that `freshenVar` won't return it as a fresh variable, which would be wrong.

In order to make the check more efficient, namely to check only the newly
added variables instead of the whole context, replace uses of the more
general `modifyContext` by more specialized `extendContext` and
`replaceContext`. The latter two know which elements are being added
and check only those.