Continuation of fix to #987 - tweak handling of let blocks in use insertion in pretty-printer
immediateChildBlockTerms was including not only some child terms but the current parent term.  That was on the assumption that its pattern for let blocks only matches at one place in the ABT for a given syntactic let block.  That's no longer true when cycle minimization is done on let blocks - the pattern matches at each step down the chain of nested ABT nodes.

However, it seems that it doesn't actually need to return the parent term anyway.  Am not sure I totally understand why, but I can't seem to find any problems with it in testing.