Convert unison to use optparse-applicative
This has a number of benefits that come at the expence of some
flexibility in how the help messages are displayed and the structure of
the commands.

* The code in main is more well typed with a lot less strings
* More logic has been moved into the options parsing such as the
number of files provided to unison transcript making runTranscripts
simpler
* The init command was missing from the help, this type of bug is less
likely to happen as the data type and structure is more visible in
Options.hs
* The limitations on how the CLI is structured lead to a more standard
way of doing things
* Free shell completion scripts with https://github.com/pcapriotti/optparse-applicative#bash-zsh-and-fish-completions

Note that optparse-applicative provides plenty of options for formatting
including native use of http://hackage.haskell.org/package/ansi-wl-pprint
so it should be possible to do _almost_ anything that's already done.

I've made a few small refactorings in Main.hs as well