Fix 1505,
wherein the sync code can't tell whether a given reference is a term or
a decl, if it's missing from the remote repo, due to remote repo being
constructed without transitive dependencies.

This only eliminates the error for `pull` users who already have the
missing dependencies in their local codebases.  It doesn't do anything
to repair the remote codebase, and arguably makes the issue less visible,
which is a good reason IMO to not merge this.