Update pull-request.create help example text
I think that this help text better reflects that `base` now follows the
`trunk` convention and that a `.prs` namespace is now encouraged.

I think that it would be good to use consistent conventions across docs,
because currently some of them are subtly different, which makes it
tougher to get started with the day-to-day development workflow.