Keep track of decompilation info while floating
This allows a much wider range of terms to be decompiled. We remember
the original term that we're about to float, and use it when the floated
combinator is decompiled.

We do this for every closed term that is being floated. Before floating,
we do a pass to close all lambda terms, so the only free variables in
them are actual recursive references. This means that every (locally)
non-recursive function has retained decompilation information.