List Endpoint: Add rootBranch & relativeTo params
Add 2 new query params to the namespace listing `/list` API:
  * `rootBranch` - a hash of the root, typically the codebase hash
  * `relativeTo` - a namespace fqn that the listing should be relative
    to.

This enables URLs that are focused into a specific namespace. For
instance if we were focused on `base.List`:

Fetch all entries in `base.List`:

  `/list?rootBranch=#mycodebasehash&relativeTo=base.List`

Fetch all `Nonempty` entries in `base.List`:

  `/list?rootBranch=#mycodebasehash&relativeTo=base.List&namespace=Nonempty`

Fetch all `Nonempty.append` entries in `base.List`:
  `/list?rootBranch=#mycodebasehash&relativeTo=base.List&namespace=Nonempty.append`

This also fixes a minor bug where the NamespaceListing were including
the root hash as the `namespaceListingHash`. This is now correctly the
hash of the returned NamespaceListing.