Start of fix for confused find-by-type (issue #2140)
There's inconsistency between hashing for V1 and V2 versions of the
codebase. The incoming search reference is constructed from the V1
version type, while the V2 types are indexed. This means we couldn't
find results for things like:

    find : [a] -> Optional a

This commit fixes the above but it's both messy and probably not
correct. It's also missing changes in various other places which have
this same issue (e.g. the "docs" command)