Add new server endpoint: /namespaces/:fqn
Add a new server endpoint for fetching the details of a namespace:
`/namespaces/:fqn`, can be used like so `/namespaces/base.List`.

This is served by the `NamespaceDetails` API which looks for a readme
(any capitalization) in the namespace and renders it.

Additionally I cleaned up a few things with regards to naming of the
previous `ListNamespace` module to match what it was returning. Down the
line the 2 APIs might be merged into a single `Namespaces` API.