Simplify the codebase server token with NanoID
Change the generated codebase server token to be a NanoID of 6
characters. A length of 6 was chosen to have a small enough chance of
collision (given the use case of the token, this could maybe be even
smaller) and minimal impact on how URLs look. NanoIDs also don't include
a lot of special characters (just _ and -), making them great for URls.