Add HTML renderer to Doc and Syntax
Mirror (as best as possible) the rendition of Doc and Syntax from the
Elm Codebase UI implementation. This is meant to eventually support a
UCM command for converting docs to html and via that build the Unison
Language Docs website.

To support this, a small Html library was also added in the style of the
Html library from Elm. It's simpler than most Haskell libraries and has
a very ergonomic API.