Doc.AsHtml avoid pre in inlineCode
pre elements are block elements and they are not allowed as children of
p (paragraph) elements. In fact the browser will automatically close an
open p if it encounters a block element in an assumption that the
developer had forgotten to close and this causes a very strange behavior
for inline code.

Also note that this "auto close" doesn't occur in client side rendering.