Fix universal eq/compare for links
It seems that optimizations cause problems together with the combination
of coercing and unsafe IO involved in reference equality. I'm not
exactly sure what goes wrong, but it causes segfaults. Defining a
noinline function for the equalities (instead of just `(==)@Reference`)
seems to fix the problem.

I got the same problem to happen for term links, so both of those use a
noinline function now. Other foreign values don't seem to be affected,
which I assume is due to them having pure primitive functions for
equality.