Fix hanging of &&/||, and elide extra parentheses
Previously, breaking a long line with an `&&` or `||` would naively
incorrectly print without indentation, like this:
```
("a long piece of text to force unison to hang the line" == "")
&&
("a long piece of text to force unison to hang the line" == "")
```
This is a syntax error. The pretty-printer now correctly hangs the line
as
```
("a long piece of text to force unison to hang the line" == "")
  && ("a long piece of text to force unison to hang the line" == "")
```

Additionally, subsequent chains of `&&` and `||` have too-restrictive
parenthesis rules, so `a && b && c` will be printed as `(a && b) && c`.
Now, the pretty-printer will search for a chain of `&&`s or `||`s and
render it without extra parens. For clarity, parentheses are kept when
`&&` and `||` are mixed.