Fix some of the TLS error handling
this is a partial fix for #2834 more will have to come. Now there is a
cert there valid for 10 years and we load it and the private key with
now problem, the call to decode the cert now is checking for
exceptions and is properly handling the case when we successfully
decode 0 certs.

Now for some reason the last section of the transcript where we try to
do handshaking, we are Segfaulting on the one of the sides of the
handshake. I haven't figured out why yet.