Fix lexing in embedded type links in docs
A doc containing an embedded term link in a doc like
`{{a doc linking to {abilityFoo}}}` would previously eagerly lex
`ability` or `type` as a keyword, tricking the parser into thinking
we're in an embedded type link, and then complain about not finding
`Foo`. This fixes the lexer to treat such cases as term links instead.

Adds a regression test transcript for making sure that `type` and
`ability` are lexed properly in type and term links in docs.