docs.as-html: move tooltips to hidden spot in dom
Rich tooltips use block elements, which, in the html spec, and in how
browsers interpret it, are not allowed inside <p> elements. This results
in the browser closing the <p> as soon as it encounters a block element,
effectively moving that block element outside of the narrative flow.

To fix this, when rendering docs as html, collect the contents of
tooltips in a separate collection (using `State` and `Writer`) from the
rest of the document, and insert them at the end of the dom in a hidden
div. Each tooltip trigger has a `data-tooltip-content-id` attribute that
can be used to render the tooltip in-place when hovering with additional
JavaScript.