Fix some compilation issues related to keeping annotations
Enclosing terms by lambdas relied upon the fact that expanding:

   \x y z -> ....

to:

   \a b c x y z -> ...

resulted in a single closed lambda term that would be caught by a lower
case. However, we sometimes now create:

  \a b c -> (\x y z -> ...) : T

and the inner lambda term is not closed. But, we don't want to re-close
the inner lambda term and float it separately, because we will
eventually just erase the annotation and produce the same series of
abstractions as the previous behavior. So, `enclose` needed to detect
this alternation of lambdas and an annotation.

Floating of let rec bindings also needed to be more intelligent in a
similar way.

I'm unsure how this led to a compilation issue, but it did, and this
seems to fix it.