Fix an issue with effect variable linearity tweaking
The rewriting was introducing empty effect lists, but this is a problem
for effect signatures in an ability, which rely on there not being any
effect ascriptions except the one for the ability being defined on the
arrow spine. E.G.

    a -> b -> c ->{A} d

is okay, but:

    a ->{} b ->{} c ->{A} d

appears like the main effect happens on the first arrow to the checker.