Make POp serialization more reliable
There was a huge list of all the ops associated with numbers used to
serialize them, but there was no way to check that every POp was in the
list (and some were actually missing). Instead, use a function from POp
to number so that the coverage checker can complain if a new op is
added. Then the list is generated from the Enum/Bounded instances.