Use original term for decompiling a top-level PAp
The new top-level floating seemed like it could cause some uninformative
decompilations if the result of floating was stored. Previously it would
have been a letrec, but now it is just the body of said letrec, and the
bound definitions have been given their own references, with no obvious
way to inspect them. So, just using the original term seems like a
better idea.

Any top-level signature is stripped off, to be slightly less verbose.