remove `clearPropagated` from `updateMetadatas`
The call to clearPropagated appears to be unnecessary: replaceFacts
traverses the intersection of the domain of (term/type)Edits and the
corresponding metadata Star3 and properly assigns the propagated
metadata values (see replaceTerms/replaceTypes). So, clearPropagated
would only have an effect if it corrected metadata for references not
in this intersection. If a reference is not in the Star3 then there is
nothing to correct. If a reference is in the Star3 but not in the
domain of termEdits then it would be erroneous to label it as not
propagated.

For example:

Suppose a codebase has only one term, B, that has the propagated
metadata. We apply a patch with the mapping A -> B. Star3.replaceFacts
has no effect since A is not in the codebase. The call to
`clearPropagated` will remove the propagated metadata from B
though. This doesn't seem correct though, B wasn't introduced by the
patch or a direct edit by the user.