Remove hacky vim auto-completion
Before Unison had LSP support for auto-completion, I added some hacky
auto-completion support to the vim plugin that called the codebase
server via curl and parsed the results with jq. The quality of the LSP
support now far exceeds this hacky completion, and I believe that the
hacky completion was accidentally disabled a while back anyway; so this
removes it.

Resolves #3847