Update nix flake for GCH 9.2.7
Resolves #4026

I was able to remove some overrides that we were using to get ormolu
0.5.0.1 to build with GHC 8.10.7. It's possible that more cruft can be
removed, but I don't understand it well enough to feel confident in
messing with it. We are no longer explicitly asking for ormolu 0.5.0.1
because I couldn't figure out how to do that, but 0.5.0.1 is the
version in the nixpkgs hash that we are currently locked on.

On my linux box both `stack build && stack test` and HLS support work.