(issue) transcript of issue with record types and user types
If a record type has a field that is a user-defined type, `view` no
longer treats the type as a record type. This becomes a big pain because
if you want to `edit` the type you have to search around for the proper
name and order of the fields to recreate what the record type definition
should look like.

**Note**: this is really more of an issue than a PR. I've added a
failing test case in a transcript but not a fix for the issue.