Add builtins for replacement in the scheme backend
These aren't actual builtins as far as the unison runtime is concerned,
but the scheme backend expects them to exist so that they can be used to
replace certain complicated, generated functions.

They are simple (un)wrapper functions for the Code and Value records. In
unison these functions are defined via serialization, but the scheme
backend can just add/remove the wrapper around the actual unison data
type representation. These are used in the implementation of
(de)serialization, so having builtins avoids a circularity.