Modify code serialization for rehashing
The serialization functions now take an extra Map that will cause some
term References to be serialized as just numbers. This can be used to
swap out the mutually-recursive references in a code SCC so that a
canonical hash for the group can be calculated. Normal serialization is
recovered by just using the empty Map.