Rework parts of primops-generated
Replaced some macro-generated stuff with a racket match. At the same
time, factored out part of the monolithing definition.

This was necessary because the dynamic code loading doesn't use the
unison function to generate an entire module, and instead uses the
functions to generate individual definitions and such. So, the decoder
for the individual pieces needs to be available.

Also filled in some missing reflection cases revealed by new tests, and
implemented some expected functionality from unison code changes.