Add missing dependency crawling for MCode
The changes to the the intermediate code formats added additional
branching constructs to MCode. But, the code that figures out the
dependencies of MCode wasn't updated to take these into account, so any
branches using these were treated as having no dependencies (due to the
default case).

I added some structure to the integration test for generating a compiled
binary to catch this case.