Restructure decompilation
The Runtime API has been changed so that 'partial' decompilation is
possible. Even if a term result is returned, a set of errors may be
yielded indicating that the entire result value could not be decompiled.

In cases with errors, the result will not be cached, although it allows
displaying a proper result.

The decompiler in the runtime now takes advantage of this, in spots that
it doesn't know how to decompile with calls to `bug`. These should
always be valid terms, but obviously won't work correctly, hence the
cache avoidance.