Tweak define-unison macro
Previously the macro would generate code like:

  (define foo
    (let ([fast (lambda ...)])
      (case-lambda ...)))

Racket seems to not be as good at optimizing this as chez was, though.
So, instead the new macro generates:

  (define-values (foo:fast foo)
    (values
      (lambda ...)
      (case-lambda ...)))

This will, worst case, allow us to directly call fast paths in the
future if necessary, which seems to be faster.

As a bonus, the macro got a little simpler.