Tweak the sandbox test for jit/interpreter comparison
For some reason, the new implementation doesn't think the existing code
should have been sandboxed for the derivations of `open`. After
considerable debugging, this seems to be due to the line:

  _ = FilePath.open

being completely floated out of the combinator before the code is sent
to the native evaluator. So in fact the sandboxing code there is
correct, the function does _not_ need to be sandboxed. However, I'm
uncertain why this discrepancy exists between the native and interpreted
paths. It is completely dead code, though, so I've modified the example
to not be dead.