Revise how native compilation works
There are, for the moment, multiple problems with producing executables.
One is that the API doesn't work when certain math libraries are
imported. The other is that the API doesn't work in a generated
executable.

So, I've turned the 'compile mode' of the standalone runtime program
just generate a module to be compiled. Then we can use an installed
`raco` to build that module into a standalone program. I don't think
there's any advantage currently to not using raco, since the best we can
do is write our own separate exe builder that still requires racket to
be installed.

In any case, the exe builder must apparently be a separate program so
that it doesn't need to import the offending math library.