Strip unison-runtime flags before evaluation
The evaluated program can ask for command line arguments, but without
intervention, it will receive all the flags from the call to the parent
program. Previously none were expected, but now we have the port.

So, as part of command line parsing, grab all the remaining args and
install them as the whole list of arguments before starting evaluation.